requirements.txt
fire==0.4.0
fonttools==4.28.3
lxml==4.7.1
numpy==1.21.4
opencv-python==4.5.4.60
pdf2docx==0.5.2
PyMuPDF==1.19.3
python-docx==0.8.11
six==1.16.0
termcolor==1.1.0
